requirements.txt
pandas==2.1.0
requests==2.31.0
psycopg2-binary==2.9.7
sqlalchemy==2.0.20
python-dotenv==1.0.0
pytest==7.4.0
pytest-cov==4.1.0
tqdm==4.66.1
